requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==7.2.2
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
selenium==4.8.2
selenium-page-factory==2.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
